Extract Code dependencies into separate files.
Fix failing webpack build due to generated html code from highlight.js
